remove libui dependancy on ui.TableString type
Use standard golang 'string' instead of requiring every external
function that manipulates the UI actually need to import the libui
toolkit. This hopefully seperates the GUI from the data the GUI
displays. The libui code actually scans through the data and
updates the cells based on what is stored in memory so this
makes this seperation clean.

Signed-off-by: Jeff Carr <jcarr@wit.com>